try to fix weird href parse error in logs
Seeing several errors similar to this:

`Argument "10330'"" isn't numeric in addition (+) at /home/dw/production/cgi-bin/DW/Controller/Search/Interests.pm line 298.`

The process of adding zero to the number with the extra trailing
punctuation results in the desired number, so it's not a bug per se,
but the error indicates that the page source isn't being parsed correctly
in some circumstances.  I couldn't reproduce the problem in Firefox,
Chrome, or Safari.

I decided to update the template to use cleaner syntax for generating
the URL, hoping that this would prevent the parse error in the
future.  If it continues to cause problems, the next step would be
to rewrite the ML string to avoid the need for string variable
interpolation in the caller.  If nothing else, removing the escaped
ampersand from the URL arguments is an improvement.