editicons: move check_form_auth back into page context
The failure message should return from page context, not
subroutine context.  This lets us eliminate the error
printing subroutine as an argument to the update_userpics
function, and eliminates a hidden BML dependency by
explicitly passing the lj_form_auth value to be checked.

This also removes a source of uninitialized value warnings
when using 'eq' to compare undefined variables.