editicons: truncate overlong input file names
I noticed in testing that excessively long image filenames
would overflow the form width, making the text to the right
of the form element unreadable.  This encloses the element
in a div with "overflow: hidden" styling.

In theory the additional file selection elements generated
in JS would benefit from the same alteration, but since we
have an issue open to change the form to support multiple
file input selection, I'm less inclined to spend more
time on this.