[#2046] print warning and skip compass step if using old version
My development laptop has a version string starting with 0.12,
so I arbitrarily chose that as the minimum viable version.

I ran into a problem, however, when checking to see what version
of Compass was installed on the admin box; I got an error message
saying that VERSION.yml was missing, so this change will likely
break the script in production until that situation is remedied.

Fixes #2046.