[#2041] fix bug with BlobStore conversion: data vs dataref
Looks like the 'store' method expects a reference to whatever
data it is given - in this case, a frozen hashref.  If it doesn't
get exactly what it expects, it croaks, and any subsequent
processing is aborted.

This fixes the reported problem in my testing, using the case
of not having enough icon slots to import all of the other
account's icons.

Fixes #2041.