[#1921] on import, suppress tag creation errors and bogus tag creation
This skips warnings for tags in @unauthorized_add when importing,
similar to what was already being done for deletion errors in
the next line.

It also addresses Azz's concern about importing bogus tags, by
not creating tags on entry import that weren't already added by
the import of all current tags, which is forced to happen first.