[#2159] adjust journal header styling to accommodate title links
For most styles, the only adjustment needed was to reapply color_page_title
to the header link.  For styles that didn't define color_page_title, I used
color_page_text instead.  (This fixed a bug in Brittle, which was using an
undefined color_page_title.)

I also removed underlining of links where appropriate.  I did my best
to blend my changes in with the existing style.

Some styles needed no adjustment.  These were in the minority.

The style that needed the most attention was Summertime.  I added new
basic styling for "#header-primary a" and moved the existing styling for
"#header a" to "#header-secondary a".

Fixes #2159.