Don't spell-check display-email; make greater errors visible
The Edit Profile page is a long way from being able to present unsaved
edits back to the user with an override checkbox (all the fields are
populated from the profile not the submission, and they're handled in
a few different ways).

Therefore I've excluded the display-email from spell-checking, but still
validated it against more serious errors. I've fixed in-passing the fact
that errors in this field would just silently not change this field,
while applying any other changes and appearing to succeed. You now get
a page with just the email error on it, if your email is very invalid.
I think this is ok because modern browsers *should* keep all your edits
when you hit the back button.