Allow email spell-checking to be overridden (#2345)
* Accept a couple more email domains; add opts{force_spelling}

* Email spelling ticky works on account create and support submit

* Tidy

This TODO was already done before I started.

Didn't mean to change initialisation of @errors.

* Ticky on change email

* Add ticky to invite.bml

* We don't need to prefix these messages

* It's not always "your" email address (gift/ban recipient)

* Ticky-support-category-edit

* Allow spell-check override in email_alias command

* Make tellafriend re-render inputs (next: ticky)

* Ticky!

* Ticky sysban

* Tickybounce: as with sysbans, this ticky is always visible

* Don't spell-check display-email; make greater errors visible

The Edit Profile page is a long way from being able to present unsaved
edits back to the user with an override checkbox (all the fields are
populated from the profile not the submission, and they're handled in
a few different ways).

Therefore I've excluded the display-email from spell-checking, but still
validated it against more serious errors. I've fixed in-passing the fact
that errors in this field would just silently not change this field,
while applying any other changes and appearing to succeed. You now get
a page with just the email error on it, if your email is very invalid.
I think this is ok because modern browsers *should* keep all your edits
when you hit the back button.

* Comment

* Tickyshop; also fix html_datetime retaining POSTed date

* Fix duplicate "selected" flag; looks like anon was the intent when logged out

* Ticky-logged-out-shop