Run Travis compile-test
This enables using Travis CI for Dreamwidth.

This is very minimal at the moment and just runs the compile test to ensure that all code compiles. We should be able to fairly trivially extend this to work with other unit tests we have.

This also updates how dependencies are tracked in a file in the `doc` directory. This should be a step towards making it easier to set up new hosts, too.