checkconfig.pl: don't re-invent version comparison
I tried to get a dev environment running during a period where the latest
version of List::Util was `1.5`. 1.5 may be larger than 1.45, but 45 is larger
than 5, so the config check failed.

Could just fix the bug and then wait to independently rediscover the next
version comparison gotcha, but I propose Let's Don't.