[#2410] Fix existing Safari quick-reply bug in some layouts
In some layouts, the quick-reply divs are preceded by floating elements.
(Example in front of me is Fluid Measure, where the entry management links are
`float: left;`.) Something about the way Safari (not Chrome, not FF) calculates
heights makes the floater overlap the outermost quick-reply div by a pixel or
two, which kicks it over to the end of the float and compresses it mightily.

This bug precedes this whole branch, and idk how many layouts it affects. It
look like telling the qrform to `clear: both;` fixes the issue without having
any other effects elsewhere.