[#2410] Stop *dramatic woodchuck* mobile zoom on input focus
If the font size of an input, select, or textarea is less than 16px, mobile
Webkit/Blink browsers will auto-zoom the viewport as soon as that element gains
focus. That's disorienting, and it also obscures part of most text fields and
makes typing jumpy.

`@media (pointer: coarse)` is an OK proxy for mobile browsers, so we can bump
the font size whenever that's true. (Might sweep up other devices that don't
woodchuck, but anything that specifically says its pointer is inaccurate would
benefit from a size bump anyway.)

I haven't yet found a way to specify "16px or higher," so 16px it is. (The CSS
`max()` function would do the trick, but it's not on a standards track anytime
soon.)

Flexible font size means we can't size the icon preview by px if we want it to
line up nicely when the icon select and the subject input are double-stacked, so
it's time to resize things with ems (after setting a base font size for all
affected elements to make sure that's reliable).