[#2434] Add qrdiv directly to page instead of attaching with jQuery
As mentioned before, throwing in a `<script>` tag at a random location in the
body content isn't the greatest. And all this script tag does is add a div with
`display: none` to the document.

Well, I know another way to add a div with `display: none` to the document: just
**add a div with `display: none` to the document!** The location doesn't matter
(since both visual renderers and readers will treat it as absent until it gets
yanked into place), so I didn't mess with the logic that just inserts it after
its first valid target.