[#2434] Un-fragment the icon browser javascript
In this case, the JS for the iconbrowser (the "old new" one for S2 pages, not
the "new new" one for the new entry page) was already being loaded the normal
way, but then the tiny fragment that was attaching it to the browse button was
separated out and squirted randomly into the body content somewhere. So, let's
move that fragment over to where it belongs. The only things it was getting from
Perl-land before getting inserted were:

- The ID of the button that triggers the browser (just hardcode it, it's never
  different in either the quickreply or the talkform).
- The saved remote state of the (small/large) (text/no-text) options (wire it
  through the form as data).