[#844] Control strip: convert tag soup function to template
There are several open bugs that require minor, straightforward, and
novice-friendly edits to the markup and styling of the control strip, all of
which have languished because it is entirely impossible to make a minor edit to
the markup of the control strip.

Using the quickreply form as a model, this commit dismantles the (totally
heinous) tag soup `$ret .=; $ret .=; etc.;` control_strip function into a
template and a relatively small data structure to drive it, which should
hopefully unblock basically anything simple that needs to be done to the control
strip. It leaves the markup structure and the "what needs to be here" logic
alone, so it shouldn't change the appearance or behavior of the control strip at
all. (There might be one or two missing `&nbsp;`s, but nothing I've been able to
spot so far.)