[#844] Control strip: Logged-out userpic is always the logged-out userpic
I dug through the blame to see whether there was any earthly reason why the
(usually empty) logged-out userpic only declares itself to be a logged-out one
if the login form DOESN'T appear. Did these hooks return interestingly
different results on LJ-alikes that didn't have the quick login form?? But it
goes all the way back to the initial import at the moment of forking, so the
explanation is lost in the mists of time.

So guess what, I TRAVELED BACK BEYOND THE MISTS OF DAMN TIME, and traced it back
to THIS COMMIT, which somehow STILL doesn't shed any light on the situation!!!
https://github.com/apparentlymart/livejournal/commit/187f71bd78e4a5a99fb84881e1f6e718c6c4af2f

Anyway, it's pointlessly illogical so I'm changing it. I'm also removing those
nbsps, because "not there" should mean not there.