[#827] S2: Don't apply global margins to control strip
To keep the control strip looking consistent, layouts that use the Tabula
Rasa/core2base HTML structure need to set global margins on `#canvas` (which has
the entire visible page content with the sole exception of the control strip)
instead of on `html body`. Most of those get that rule from core2base's
`print_default_stylesheet()`, so this commit doesn't have to touch most of the
affected layouts.

To test this change, go to the "Base Layouts" section of /customize and compare
the results for the following layouts:

* Lefty
* Leftovers
* Seamless
* Paper Me
* Motion
* Librarian's Dream
* Snakes & Boxes
* Ciel

Other layouts can be affected under the right circumstances, but those are the
ones that are wonky by default, which makes it easier to see the fixed behavior.