The previous change to error_ml dropped any message arguments for LJ::Lang::ml on the floor, because the expected inputs weren't well documented.
This changes error_ml to accept an input format similar to success_ml, where
the message arguments are in the first hashref, and any page variables are
in a subsequent hashref.

The two places where page variables had begun being passed into error_ml are
updated to pass in undef for the message arguments, so everything should line
up correctly.

Fixes #2430.