[#2434] Don't insert jQuery script tags into main body content (#2442)
* [#2434] Add qrdiv directly to page instead of attaching with jQuery

As mentioned before, throwing in a `<script>` tag at a random location in the
body content isn't the greatest. And all this script tag does is add a div with
`display: none` to the document.

Well, I know another way to add a div with `display: none` to the document: just
**add a div with `display: none` to the document!** The location doesn't matter
(since both visual renderers and readers will treat it as absent until it gets
yanked into place), so I didn't mess with the logic that just inserts it after
its first valid target.

* [#2434] Defer quote button javascript with need_res

Previously, the jQuery function that creates and arms the quote button for the
reply forms got squirted somewhere into the middle of the body content as a
free-floating script tag. This commit moves it into a normal script resource
that gets loaded like other, similar script resources, using need_res.

The generated JS was getting two things from Perl-land:

- The ID of the textarea to paste into. There are only two options (the one used
  in the quickreply and the one used in the talkform), so we'll just figure it
  out once we get there.
- The (possibly localized) error/help text if you click the quote button without
  a selection. Easy to embed that in a data attribute.

* [#2434] Un-fragment the icon browser javascript

In this case, the JS for the iconbrowser (the "old new" one for S2 pages, not
the "new new" one for the new entry page) was already being loaded the normal
way, but then the tiny fragment that was attaching it to the browse button was
separated out and squirted randomly into the body content somewhere. So, let's
move that fragment over to where it belongs. The only things it was getting from
Perl-land before getting inserted were:

- The ID of the button that triggers the browser (just hardcode it, it's never
  different in either the quickreply or the talkform).
- The saved remote state of the (small/large) (text/no-text) options (wire it
  through the form as data).

* [#2434] Normalize deferred setup to `jQuery(function($){...})`

`jQuery(function($){...})` is a terser synonym for
`jQuery(document).ready(function($){...})`. A quick grep of the code shows it's
more common here, so let's tidy these files up while I'm in here.

* Cleanup: Adjust height of quickreply form elements

The original size was a little too small, and could sometimes cut off part of
the text. This bumps it up another .2em, and also extracts it into a variable so
it's easier to update all the spots if I still didn't get it quite right.

* Cleanup: Use qq{} to avoid ugly concats or quote escapes

* Tidy Talk.pm/Web.pm