[#2437] Extract talkform into template
This commit is intended as a no-op, and it specifically **doesn't** address the
fact that the talkform is awful garbage. That comes later. The only changes in
behavior are a few fixed bugs where the intended behavior was reasonably clear
but the implementation just didn't _do_ it. (Hiding the "otheruser" row but not
hiding lj_more, hiding the openID login fields for a case that was exactly
equivalent to another case where it showed them, blatantly un-closed or
over-closed HTML tags, etc.)