Remove SSL shenanigans
This diff removes IS_SSL, USE_SSL, USE_HTTPS_EVERYWHERE,
ALL_TRAFFIC_IS_SSL, SSLROOT, SSL*PREFIX, and a bunch of other machinery
that was responsible for SSL management.

This used to be necessary when we had some parts of the site that were
SSL and some weren't, but that's no longer true. These days, the
expectation is that all web traffic is encrypted and that is true on
Dreamwidth and should be true on our hacks and anybody else using our
code, so we can truly simplify the code.

The recommended way for running DW is behind something like Cloudflare
or AWS CloudFront or another proxy service that provides cheap, fast SSL
at scale. If you happen to be running it yourself, you can do the same
with Nginx, Varnish, or your reverse proxy of choice.