Remove public key/PGP functionality
I looked at how many people have submitted public keys. The number is
20. This feature is relatively unused and can go away.

If people want to link to their public keys they can put them in their
profile and/or link to one of the web sites where you can validate and
claim your identity, which is far better solution than trusting the
public key off of Dreamwidth or whatnot. We're a blogging service, we
don't really have the kind of system/promises to guarantee security or
strong identity.