[#2432] Allow users to opt-in to responsive indent in Tabula Rasa (and children)
This enables the comment indent S2 properties in core2base, and adds some basic
responsive indent CSS (which is only added to the stylesheet if a user switches
the indent style to "Responsive").

Basic means basic -- this uses a 10px indent on mobile, and after a depth of 8
it bails out and starts exposing the textual depth indicators. On desktop, it
uses 20px (only a smidge tighter than the old static 25px), and bails to textual
indicators at 180 deep.