[#2432] Enable advanced responsive comment threading in Practicality (opt-in)
This is sort of the proving ground, and its implementation will probably change
as we get feedback on it. For one thing, implementing this in a style could
REALLLLLY stand to be easier; the stuff I added in core2 makes this _possible,_
but it doesn't make it _simple_ yet, and a style author can do way too much
wacky stuff that risks becoming unsupportable. So be on the lookout for a more
prescriptive way to handle this. Welcome to Beta Zone.

Denise and Momiji said Practicality was the most popular of the Rasa-scions, so
it gets to be the guinea pig.

Alyndra had the good idea of making thread lines go _behind_ comments once you
hit the limit, so the eye can see a square wave sort of thing going on as the
depth goes in and out.

Props also to Highlander_II for pointing out that Reddit does something much
like this, which helped validate that I might not be on the wrong track.

The every-fifth-one-is-different thing is the one part I'm gonna straight-up
brag about on here, IMO that was a stroke of brilliance. 🌻