[#2485] Fix some mobile jank in Tabula Rasa/core2base and in Practicality
The userpic float -> absolute stuff in Practicality might seem a little odd, so
here's what's up with that:

- We want `overflow-x: auto` on content containers for entries and comments,
  because nothing posted by a rando should EVER be allowed to trash your flist
  or entry page layouts into full-page sideways slop.
- But this can have unexpected effects on containers that try to send something
  outside the box with relative positioning, which I guess includes negative
  margins. The astronaut content gets decapitated as soon as it sticks its head
  outside the `overflow-x: auto` container.

The way around that is to do something like this:

- outer container with `position: relative`
    - inner container with `overflow-x: auto`
        - child element with `position: absolute`

The "containing block" of an `absolute` box is its nearest `relative` ancestor.
That means the `absolute` items now escape the `overflow-x: auto` container,
because it is NOT THEIR REAL MOM!!! and they are GOING TO GRANDMA'S!!!, which
prevents them from getting clipped.

Then you still need to have the text float around _something,_ which is what
that ::before pseudo-element with float is all about.