[#2437] Extract talkform into template (#2479)
* Cleanup: add missing copyright/author header to controlstrip.tt

* Web.pm: Extract userpic lister into a function

* Improve subjecticon implementation (breaks talkform, fixing in next commit)

- Use more comprehensible function names.
- Don't require callers to constantly pass around that hashref.
- Abstract the print functions enough that we could easily replace the
  implementation with big emoji or something.
- Add 'none' as an option in the index, so you don't need a special function to
  print the "none image".
- Cache the (hardcoded) results of get_subjecticons in a closed-over var,
  instead of re-calculating every time.
- Actually document it.

* [#2437] Extract talkform into template

This commit is intended as a no-op, and it specifically **doesn't** address the
fact that the talkform is awful garbage. That comes later. The only changes in
behavior are a few fixed bugs where the intended behavior was reasonably clear
but the implementation just didn't _do_ it. (Hiding the "otheruser" row but not
hiding lj_more, hiding the openID login fields for a case that was exactly
equivalent to another case where it showed them, blatantly un-closed or
over-closed HTML tags, etc.)

* [#2437] OpenID is always enabled

Changed upstream while the branch was in progress.

* Talkform: Fix bug when posting as a different user than the logged-in one

* Run tidy

* Style adjustments from PR review

* Simplify handling of subjecticons in talkform

Whenever practical, pass around IDs instead of hashrefs.

* Re-tidy Talk.pm