Styles: Wrong keyword value in image fix
I used a Firefox-ism that isn't broadly supported! When an image with height and
width attributes gets shrunk to fit, this can cause some white space above and
below it in other browsers. `height: auto;` is equivalent to what I meant here.
(It overrides the height attribute, but sets it to match the native aspect ratio
for the width we ended up using. 99.5% of the time this is correct; the other
.5%, you can override it with an inline `min-height`.)