Fix thick-bottomed userpic boxes (#2490)
* Fix thick-bottomed userpic boxes

A normal <a> is an inline element whose height is derived from font metrics and
line-height, rather than from the height of the box-things in it. The <a> around
a userpic thus adds some extra height to the surrounding div, making styling
difficult and often adding visual glitches. If you make it a block element
and remove the line-height, it will fit snugly to its <img> contents instead.

* Let's get the one in the controlstrip too!