Fix quickreply appearing in center of the page on sitescheme/lightscheme
This left/right margin: auto was doing almost nothing on journal-styled
pages, but was doing the wrong thing on site/light pages (because they don't use
the same structure of nested divs). This should make the quickreply start at
the same indent level as the comment being replied to.

Conflicts:
	htdocs/scss/components/quick-reply.scss