userpicfactory.js: Revert to using an on-load handler
SOMETHING breaks if you try to do this setup as soon as the DOM is ready. Either
it's relying on image dimensions that aren't available until the image content
is loaded (been there), or there's some kind of existing race condition that
isn't exposed until this code gets faster. Hard to say, tbh.