Implement @foo user mentions
As per suggestion: https://dw-suggestions.dreamwidth.org/1527807.html

This makes the usage "@foo" and "@foo.BAR" work everywhere, where BAR is
one of the external sites. Such as "@foo.twitter" or "@foo.lj". This is
what we already do in Markdown so it's easy to expand it to the rest of
our HTML cleaning.