Quickreply: Keep bottom buttons from stretching
The default value of align-items is 'stretch', which fills vertical (well,
cross-axis) space to match the tallest item in the container. That's basically
never what we want in this bottom button row.

If something is way too tall, there'll still be undesirable space between
the buttons. But the buttons themselves won't be clownshoes.