Quickreply: RIP quick-er reply
With spellcheck gone, the only two differences between the forms were
the preview button and the quote button, and I don't see a good reason
to exclude either of them from the reading page / recent page. This
option might come back once we unify the talkform and the quickreplies
since we'll be bringing in user-switching controls and the like, so I'm
holding off on removing that from the function's options hash for now.