[#2535] Let quick-reply stretch its legs if there's room on the left
Standard comment indentation can shrink comment boxes down to slivers, and the
initial version of the responsive quick-reply was constrained to the comment
box. This leaves the form feeling VERY CRAMPED. (It used to just overflow
rightwards, because tables gonna do what tables gonna do.)

But if the box shrank because of indentation, then by definition there's a TON
of unused room off on the left. This commit does some simple measurements
whenever we summon the quickreply into place, and if staying in the box would
make it too narrow, we expand it (and give it a shove to the left to make sure
it overflows in the correct direction). We also take a peek to see what
background color it probably has when its inside the box, and apply that
background color to the overflowed part so it looks less like glitch art. (The
adjustments to padding are to make sure this looks nice.)

What's "too narrow?" If the viewport is big and there's plenty of room, 640px
seems like a nice minimum. Once you take the padding into account, that leaves
the form wide enough for about 45 characters of 13px Courier (a common default
for textareas). If viewport is small, we're probably space-constrained in the
first place and actually never want the comment form to shrink below where it
would be in a top-level comment. So in that case, we find the first comment on
the page (which is _always_ a top-level reply to the original post, because of
how DW collapses comment threads once pagination kicks in) and measure how much
space a comment would have up there.