QR: Move quote button up by subject
The buttons at the bottom should be things you do when you're _done_ writing
your comment. Quote is a formatting button, useful _while_ writing your comment.
So it should be up higher by the comment-writing controls, not the
comment-ending controls.

As part of this, I'm removing it from the tab index, so keyboard users don't
have to tab through it to get from subject to body. Note this pattern also on
GitHub's formatting buttons. I'm also removing the `display: none`; we don't
need to hide it for no-js users, because the quick-reply form can only be
summoned with JavaScript enabled anyway.

A future design will likely include a row of formatting controls between the
subject and body, like you can see here on GitHub when commenting or opening an
issue. When that happens, quote will join those. In the meantime, putting it
beside subject keeps it nicely out of the way.

TBH I already wanted to do this during the initial redesign, but I couldn't
stomach what happened when the qrform got squished down below like 200px. But
now it's kicking out to the left to preserve its size when it's deep in a
comment thread, so we're free!