Better Markdown, maybe
This moves Markdown and @-user conversions down into the HTML cleaner,
but only when we're processing text nodes and only when we're not in the
middle of HTML.

This means that for a user that is only using Markdown, things will work
great and everything will be "owned" by Markdown.

While using Markdown, you can also "drop to HTML" by opening any tag.
This will suspend Markdown processing and @-user mentions until you
close the tag. This enables you to drop in a textarea or whatever and
not have to stress about it.

For HTML users, we of course disable both Markdown and @-user mentions.

This also makes syndicated and imported content disable both of the
above to make sure we don't frotz them inappropriately.

I'm not convinced this is right yet, I slapped it together this
afternoon after some thinking and would like to write some more tests
before committing it.