[#2545] QR: Use `nouserpic_sitescheme` as data-url when there's no default icon
For users with no icons, we display the faded-out DW logo in place of the icon
preview in the quickreply. We also do that for users with no default icon, but
if they switch to a real icon and then switch _back_ to default, we replace it
with an empty glitched img tag, because the `data-url` attribute for `(default)`
is an empty string.

This commit checks whether the user's `defaultpicurl` is empty, and if so, it
falls back to the URL of the same `nouserpic_sitescheme` placeholder icon.