Talkform: always include quote button
I ported this logic over from the original talkform, because I was trying to be
thorough:

https://github.com/dreamwidth/dw-free/pull/2479/files#diff-339927c4adfb5b1e94ddf9bb2910209eL2158

But it made no sense to me at the time, and now that I've had time to reflect,
I'm pretty sure it's nonsense. There's no good reason to exclude the quote
button for anonymous users, and there's not even a great reason to exclude it on
error-fixes or captchas -- yeah, it's not very useful, but why bother
reconfiguring the form? Anyway, bye.