[#2549] Return to journal pages on logout
This restores the ability to click the log out button in the controlstrip and
get returned to the same page, much beloved by RPers and other multi-account
users.

Unlike the original implementation, I'm using a hidden form input instead of a
URL parameter (`?ret=1`) -- since this is a POST action, it seems like the
arguments should all arrive via POST data.

This also fixes an inconsequential copypasta glitch in the controller's name.