[#2427] ljlib: include lj_base.css in Foundation pages
Because why not. This CSS file includes some basic styles for objects that
appear in entry/reply/icons pages; the Foundation pages haven't needed any of
that in the past, but they will eventually. As far as I can tell, the rules here
don't have any meaningful conflicts with existing rules in the Foundation site
skins.