Add all S2 javascript includes to the 'all' group, instead of 'jquery'
It's currently impossible for S2 pages to end up in the `'default'` (actually
now legacy) resource group for JavaScript. However, they might be in either the
`'jquery'` or `'foundation'` groups, depending on whether we're using sitepages.
So we need to make sure any JS we include will be available in both groups.