[#2427] Include S2 javascripts in the 'all' group, not 'jquery'
It's currently impossible for S2 pages to end up in the `'default'` (read:
legacy) resource group. However, they might be in either the `'jquery'` or
`'foundation'` group, depending on whether you've enabled the beta. So we need
to make sure any JS we include will be available in both groups, which is what
the `'all'` group will do.

Known issue: The icon browser is broken in journal pages, because this commit
doesn't touch init_iconbrowser_js.