[#2555] Foundation core: Update Foundation to 5.5.3
I'm preserving our changes to _settings.scss and _foundation.scss, but also
trying to keep it easy to compare our edits to the base files. Not that it
matters, since there's never going to be another release of Foundation 5 and
upgrading to 6 sounds like some real live heck!!

The only major feature motivating this update is the ability to get off the
ancient and deprecated sass and Compass gems we've been stuck on:

- Foundation 5.4 ONLY works with Ruby Sass 3.2 and Compass 0.12.
- Foundation 5.5 ONLY works with Sass 3.4 and higher and Compass 1.0 and higher.

Of course, even the newest version of Compass is now deprecated abandonware. Sigh.