Don't hide logout button on /logout page
I get where this was coming from, but the result is that the page looks
glitched -- there's no other circumstance where the site header changes
like that, and it also throws away the username link, which is a useful
control in its own right. The benefit of consistency is greater than
the drawback of redundancy, in this case.