Foundation futureproofing: Change `$small` references to `$medium-up`
...because that's apparently what it currently means. o_o WOW.

In our current and previous versions of Foundation, there's typo in the
definitions for the legacy size class variables, and our nav styles were relying
on the typo. A future version fixes the typo, which will of course blow
everything up in our site headers. Anyway, anything that asks for `$small` in
its media query is currently getting the `$medium-up` query, so it should just
ask for that by its real name.