Talkform: Fix preview-submits-for-real bug
This was masked by my original typo on that JS that disables the form buttons on
submit. Once the typo was fixed, the preview button started insta-posting. 😱

A quick investigation revealed that some part of our Perl-land form handling
causes disabled submit buttons to have a falsy value. And it turns out the
quickreply already had a workaround for that! It just didn't have any comments
explaining _why_ it was doing that workaround, so I didn't know to reproduce it.

So I fixed the talkform bug the same way, and added some comments explaining the
workarounds.

Note that if you have JS disabled, the talkform preview button works as expected
-- it starts out with the correct input name. If JS is enabled enough to dummy
out the name, it's also enabled enough to send the correct value in the hidden
input.